[XAM] Split content device exports to a new file.
