+++ src/xenia/kernel/xam/xam_module_export_groups.inc
@@ -12,6 +12,7 @@
 
 XE_MODULE_EXPORT_GROUP(xam, Avatar)
 XE_MODULE_EXPORT_GROUP(xam, Content)
+XE_MODULE_EXPORT_GROUP(xam, ContentDevice)
 XE_MODULE_EXPORT_GROUP(xam, Enum)
 XE_MODULE_EXPORT_GROUP(xam, Info)
 XE_MODULE_EXPORT_GROUP(xam, Input)
